+++ package.json
@@ -26,6 +26,8 @@
     "react-redux": "^7.2.1",
     "react-router": "^5.2.0",
     "redux": "^4.0.5",
+    "redux-devtools-extension": "^2.13.8",
+    "redux-thunk": "^2.3.0",
     "sass": "^1.26.10",
     "sass-loader": "^9.0.3",
     "url-loader": "^4.1.0",
