+++ src/components/Card/Card.scss
@@ -1,3 +1,5 @@
+@import '../../styles/variables.scss';
+
 .card {
   margin-bottom: 50px;
   color: $color-darker;
@@ -45,7 +47,7 @@
     height: 40px;
     padding: 5px;
     border-radius: 20px;
-    background: url("./assets/kebab_menu_inverted.png");
+    background: url("../../assets/kebab_menu_inverted.png");
     background-color: $color-border;
     background-size: 20px;
     background-repeat: no-repeat;
