+++ src/components/Form/Form.scss
@@ -1,3 +1,5 @@
+@import '../../styles/variables.scss';
+
 .form-section {
   margin-bottom: 25px;
 }
@@ -31,7 +33,7 @@
       color: rgba($color: #000000, $alpha: 0.0);
       display: block;
       opacity: 1;
-      background: url('./assets/calendar.png') no-repeat;
+      background: url('../../assets/calendar.png') no-repeat;
       width: 25px;
       height: 25px;
 
@@ -44,7 +46,7 @@
     -moz-appearance: none;
     appearance: none;
 
-    background: url('./assets/arrow.png') 95% / 5% no-repeat $color-border;
+    background: url('../../assets/arrow.png') 95% / 5% no-repeat $color-border;
 
     @include hover;
   }
