+++ src/components/Modals/Modal.scss
@@ -1,3 +1,5 @@
+@import '../../styles/variables.scss';
+
 .modal {
   @extend .font_thin;
