+++ src/components/SorterDropdown/SorterDropdown.scss
@@ -1,3 +1,5 @@
+@import '../../styles/variables.scss';
+
 .dropdown {
   position: relative;
   display: inline-block;
