+++ src/index.js
@@ -1,8 +1,8 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import Landing from "./views/Landing.View";
-import ErrorBoundary from "./components/ErrorBoundary";
-import VODCollection from "./views/VODCollection.View";
+import Landing from "./views/Landing/Landing";
+import ErrorBoundary from "./components/ErrorBoundary/ErrorBoundary";
+import VODCollection from "./views/VODCollection/VODCollection";
 import { assetsPath, VODList } from "./models/VODCollection.DataModel";
 
 import "./index.scss";
