+++ src/index.scss
@@ -1,14 +1,6 @@
+@import './styles/variables.scss';
+
 //==== Components
 
 @import "./styles/base.scss";
 @import "./styles/header.scss";
-@import "./styles/card.scss";
-@import "./styles/sorterDropdown.scss";
-@import "./styles/modal.scss";
-@import "./styles/form.scss";
-@import "./styles/searchForm.scss";
-
-//===== Views
-@import "./styles/views/Landing.scss";
-@import "./styles/views/VODCollection.scss";
-@import "./styles/views/DetailsPage.scss";
