--- src/styles/views/Landing.scss
@@ -1,33 +1,0 @@
-@mixin bluredBackground() {
-  &:before {
-    content: "";
-    width: 100%;
-    height: 100%;
-    position: absolute;
-    z-index: -99;
-    background: url("./assets/background.jpg") repeat;
-    background-size: cover;
-    filter: blur(5px) brightness(0.5);
-
-    @content;
-  }
-}
-
-.landing {
-  height: 100vh;
-  color: $color-accent;
-
-  @include bluredBackground;
-
-  main {
-    position: relative;
-    top: calc(50% - #{$search-height});
-  }
-
-  &--search-title {
-    font-size: 48px;
-    text-transform: uppercase;
-    color: $color-white;
-    padding-bottom: 20px;
-  }
-}
