+++ src/views/VODCollection/VODCollection.scss
@@ -1,3 +1,5 @@
+@import '../../styles/variables.scss';
+
 .VODCollection {
   color: $color-white;
